feat: added generateid func

+++ pkg/notp/statemachines/packets/packet.go
@@ -40,7 +40,7 @@ const (
 	CompletedDataStreamValue = uint32(4)
 
 	// FlowIDValue represents the flow ID.
-	FlowIDValue = uint32(10)
+	FlowIDValue = uint16(10)
 
 	// StartFlowMessage represents the notification of the flow.
 	StartFlowMessage = uint16(100)
